data saved correctly and module for loaders
