+++ author.rb
@@ -15,12 +15,12 @@ class Author
 
   def to_json(*args)
     {
-      JSON.create_id => self.class.name, 'props' => [@id, @title, @color]
+      JSON.create_id => self.class.name, 'props' => [@id, @first_name, @last_name]
     }.to_json(*args)
   end
 
   def self.json_create(object)
-    author = new(title: object['props'][1], color: object['props'][2])
+    author = new(first_name: object['props'][1], last_name: object['props'][2])
     author.id = object['props'][0]
     author
   end
