+++ books.json
@@ -1,0 +1,1 @@
+[{"json_class":"Book","a":[40,{"json_class":"Label","a":[76,"title","color"]},"publisher","good","yes","dd-mm-yyyy"]}]
