+++ games.json
@@ -1,0 +1,1 @@
+[{"json_class":"Game","props":[23,{"json_class":"Author","props":[12,"axel","soler"]},"hoy","n","y","ayer"]}]
