+++ labels.json
@@ -1,0 +1,1 @@
+[{"json_class":"Label","a":[76,"title","color"]}]
